PP-6002 Add CreateTokenRequest type with validation
Following our conventions for using '@Valid' annotations this adds a new
CreateTokenRequest to replace the use of JsonNode. It copies the validation
rules from before. To avoid deviating from default behaviour the error
response code for failed validation will be 422 (rather than 400) and the
error message changes slightly but no detail is lost; tests updated to reflect this.